@@ -2,10 +2,14 @@ name: google_maps_flutter
 description: A Flutter plugin for integrating Google Maps in iOS and Android applications.
 author: Flutter Team <flutter-dev@googlegroups.com>
 homepage: https://github.com/flutter/plugins/tree/master/packages/google_maps_flutter
-version: 0.0.1
+version: 0.1.0
 
 dependencies:
   flutter:
+    sdk: flutter
+
+dev_dependencies:
+  flutter_test:
     sdk: flutter
 
 flutter:
@@ -14,6 +18,7 @@ flutter:
     iosPrefix: FLT
     pluginClass: GoogleMapsPlugin
 
+
 environment:
   sdk: ">=2.0.0-dev.47.0 <3.0.0"
-  flutter: ">=0.3.0 <2.0.0"
+  flutter: ">=0.11.9 <2.0.0"
