@@ -2,6 +2,7 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
+import 'package:flutter/services.dart';
 import 'package:meta/meta.dart';
 import 'package:flutter/widgets.dart';
 
@@ -19,8 +20,8 @@ String defaultNameExtractor(RouteSettings settings) => settings.name;
 /// A [NavigatorObserver] that sends events to Firebase Analytics when the
 /// currently active [PageRoute] changes.
 ///
-/// When a route is pushed or poped, [nameExtractor] is used to extract a name
-/// from [RouteSettings] of the now active route and that name is send to
+/// When a route is pushed or popped, [nameExtractor] is used to extract a name
+/// from [RouteSettings] of the now active route and that name is sent to
 /// Firebase.
 ///
 /// The following operations will result in sending a screen view event:
@@ -49,18 +50,39 @@ String defaultNameExtractor(RouteSettings settings) => settings.name;
 /// [PageRouteAware] and subscribing it to [FirebaseAnalyticsObserver]. See the
 /// [PageRouteObserver] docs for an example.
 class FirebaseAnalyticsObserver extends RouteObserver<PageRoute<dynamic>> {
+  /// Creates a [NavigatorObserver] that sends events to [FirebaseAnalytics].
+  ///
+  /// When a route is pushed or popped, [nameExtractor] is used to extract a
+  /// name from [RouteSettings] of the now active route and that name is sent to
+  /// Firebase. Defaults to `defaultNameExtractor`.
+  ///
+  /// If a [PlatformException] is thrown while the observer attempts to send the
+  /// active route to [analytics], `onError` will be called with the
+  /// exception. If `onError` is omitted, the exception will be printed using
+  /// `debugPrint()`.
   FirebaseAnalyticsObserver({
     @required this.analytics,
     this.nameExtractor = defaultNameExtractor,
-  });
+    Function(PlatformException error) onError,
+  }) : _onError = onError;
 
   final FirebaseAnalytics analytics;
   final ScreenNameExtractor nameExtractor;
+  final void Function(PlatformException error) _onError;
 
   void _sendScreenView(PageRoute<dynamic> route) {
     final String screenName = nameExtractor(route.settings);
     if (screenName != null) {
-      analytics.setCurrentScreen(screenName: screenName);
+      analytics.setCurrentScreen(screenName: screenName).catchError(
+        (Object error) {
+          if (_onError == null) {
+            debugPrint('$FirebaseAnalyticsObserver: $error');
+          } else {
+            _onError(error);
+          }
+        },
+        test: (Object error) => error is PlatformException,
+      );
     }
   }
 
