@@ -6,14 +6,34 @@ package io.flutter.plugins.firebaseauth;
 
 import android.net.Uri;
 import android.support.annotation.NonNull;
+import android.support.annotation.Nullable;
 import android.util.SparseArray;
 import com.google.android.gms.tasks.OnCompleteListener;
 import com.google.android.gms.tasks.Task;
 import com.google.firebase.FirebaseApiNotAvailableException;
 import com.google.firebase.FirebaseApp;
 import com.google.firebase.FirebaseException;
+import com.google.firebase.FirebaseNetworkException;
 import com.google.firebase.FirebaseTooManyRequestsException;
-import com.google.firebase.auth.*;
+import com.google.firebase.auth.AuthCredential;
+import com.google.firebase.auth.AuthResult;
+import com.google.firebase.auth.EmailAuthProvider;
+import com.google.firebase.auth.FacebookAuthProvider;
+import com.google.firebase.auth.FirebaseAuth;
+import com.google.firebase.auth.FirebaseAuth.AuthStateListener;
+import com.google.firebase.auth.FirebaseAuthException;
+import com.google.firebase.auth.FirebaseAuthInvalidCredentialsException;
+import com.google.firebase.auth.FirebaseUser;
+import com.google.firebase.auth.GetTokenResult;
+import com.google.firebase.auth.GithubAuthProvider;
+import com.google.firebase.auth.GoogleAuthProvider;
+import com.google.firebase.auth.PhoneAuthCredential;
+import com.google.firebase.auth.PhoneAuthProvider;
+import com.google.firebase.auth.PhoneAuthProvider.ForceResendingToken;
+import com.google.firebase.auth.SignInMethodQueryResult;
+import com.google.firebase.auth.TwitterAuthProvider;
+import com.google.firebase.auth.UserInfo;
+import com.google.firebase.auth.UserProfileChangeRequest;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -23,22 +43,19 @@ import java.util.ArrayList;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.List;
+import java.util.Locale;
 import java.util.Map;
 import java.util.concurrent.TimeUnit;
 
 /** Flutter plugin for Firebase Auth. */
 public class FirebaseAuthPlugin implements MethodCallHandler {
   private final PluginRegistry.Registrar registrar;
-  private final SparseArray<FirebaseAuth.AuthStateListener> authStateListeners =
-      new SparseArray<>();
-  private final SparseArray<PhoneAuthProvider.ForceResendingToken> forceResendingTokens =
-      new SparseArray<>();
+  private final SparseArray<AuthStateListener> authStateListeners = new SparseArray<>();
+  private final SparseArray<ForceResendingToken> forceResendingTokens = new SparseArray<>();
   private final MethodChannel channel;
 
   // Handles are ints used as indexes into the sparse array of active observers
   private int nextHandle = 0;
-
-  private static final String ERROR_REASON_EXCEPTION = "exception";
 
   public static void registerWith(PluginRegistry.Registrar registrar) {
     MethodChannel channel =
@@ -53,7 +70,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private FirebaseAuth getAuth(MethodCall call) {
-    Map<String, Object> arguments = (Map<String, Object>) call.arguments;
+    Map<String, Object> arguments = call.arguments();
     String appName = (String) arguments.get("app");
     FirebaseApp app = FirebaseApp.getInstance(appName);
     return FirebaseAuth.getInstance(app);
@@ -71,8 +88,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "createUserWithEmailAndPassword":
         handleCreateUserWithEmailAndPassword(call, result, getAuth(call));
         break;
-      case "fetchProvidersForEmail":
-        handleFetchProvidersForEmail(call, result, getAuth(call));
+      case "fetchSignInMethodsForEmail":
+        handleFetchSignInMethodsForEmail(call, result, getAuth(call));
         break;
       case "sendPasswordResetEmail":
         handleSendPasswordResetEmail(call, result, getAuth(call));
@@ -86,23 +103,11 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "delete":
         handleDelete(call, result, getAuth(call));
         break;
-      case "signInWithEmailAndPassword":
-        handleSignInWithEmailAndPassword(call, result, getAuth(call));
-        break;
-      case "signInWithGoogle":
-        handleSignInWithGoogle(call, result, getAuth(call));
+      case "signInWithCredential":
+        handleSignInWithCredential(call, result, getAuth(call));
         break;
       case "signInWithCustomToken":
         handleSignInWithCustomToken(call, result, getAuth(call));
-        break;
-      case "signInWithFacebook":
-        handleSignInWithFacebook(call, result, getAuth(call));
-        break;
-      case "signInWithTwitter":
-        handleSignInWithTwitter(call, result, getAuth(call));
-        break;
-      case "signInWithGithub":
-        handleSignInWithGithub(call, result, getAuth(call));
         break;
       case "signOut":
         handleSignOut(call, result, getAuth(call));
@@ -110,20 +115,14 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       case "getIdToken":
         handleGetToken(call, result, getAuth(call));
         break;
-      case "linkWithEmailAndPassword":
+      case "reauthenticateWithCredential":
+        handleReauthenticateWithCredential(call, result, getAuth(call));
+        break;
+      case "linkWithCredential":
         handleLinkWithEmailAndPassword(call, result, getAuth(call));
         break;
-      case "linkWithGoogleCredential":
-        handleLinkWithGoogleCredential(call, result, getAuth(call));
-        break;
-      case "linkWithFacebookCredential":
-        handleLinkWithFacebookCredential(call, result, getAuth(call));
-        break;
-      case "linkWithTwitterCredential":
-        handleLinkWithTwitterCredential(call, result, getAuth(call));
-        break;
-      case "linkWithGithubCredential":
-        handleLinkWithGithubCredential(call, result, getAuth(call));
+      case "unlinkFromProvider":
+        handleUnlinkFromProvider(call, result, getAuth(call));
         break;
       case "updateEmail":
         handleUpdateEmail(call, result, getAuth(call));
@@ -157,8 +156,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
 
   private void handleSignInWithPhoneNumber(
       MethodCall call, Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    Map<String, String> arguments = call.arguments();
     String verificationId = arguments.get("verificationId");
     String smsCode = arguments.get("smsCode");
 
@@ -171,10 +169,10 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
 
   private void handleVerifyPhoneNumber(
       MethodCall call, Result result, final FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    final int handle = call.argument("handle");
-    String phoneNumber = call.argument("phoneNumber");
-    int timeout = call.argument("timeout");
+    Map<String, Object> arguments = call.arguments();
+    final int handle = (int) arguments.get("handle");
+    String phoneNumber = (String) arguments.get("phoneNumber");
+    int timeout = (int) arguments.get("timeout");
 
     PhoneAuthProvider.OnVerificationStateChangedCallbacks verificationCallbacks =
         new PhoneAuthProvider.OnVerificationStateChangedCallbacks() {
@@ -223,7 +221,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         };
 
     if (call.argument("forceResendingToken") != null) {
-      int forceResendingTokenKey = call.argument("forceResendingToken");
+      int forceResendingTokenKey = (int) arguments.get("forceResendingToken");
       PhoneAuthProvider.ForceResendingToken forceResendingToken =
           forceResendingTokens.get(forceResendingTokenKey);
       PhoneAuthProvider.getInstance()
@@ -248,9 +246,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private Map<String, Object> getVerifyPhoneNumberExceptionMap(FirebaseException e) {
-    Map<String, Object> exceptionMap = new HashMap<>();
     String errorCode = "verifyPhoneNumberError";
-
     if (e instanceof FirebaseAuthInvalidCredentialsException) {
       errorCode = "invalidCredential";
     } else if (e instanceof FirebaseAuthException) {
@@ -260,6 +256,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     } else if (e instanceof FirebaseApiNotAvailableException) {
       errorCode = "apiNotAvailable";
     }
+
+    Map<String, Object> exceptionMap = new HashMap<>();
     exceptionMap.put("code", errorCode);
     exceptionMap.put("message", e.getMessage());
     return exceptionMap;
@@ -267,8 +265,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
 
   private void handleLinkWithEmailAndPassword(
       MethodCall call, Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+    Map<String, String> arguments = call.arguments();
     String email = arguments.get("email");
     String password = arguments.get("password");
 
@@ -279,7 +276,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleCurrentUser(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+  private void handleCurrentUser(
+      @SuppressWarnings("unused") MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     FirebaseUser user = firebaseAuth.getCurrentUser();
     if (user == null) {
       result.success(null);
@@ -290,14 +288,13 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private void handleSignInAnonymously(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+      @SuppressWarnings("unused") MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth.signInAnonymously().addOnCompleteListener(new SignInCompleteListener(result));
   }
 
   private void handleCreateUserWithEmailAndPassword(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
     String email = arguments.get("email");
     String password = arguments.get("password");
 
@@ -306,21 +303,19 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleFetchProvidersForEmail(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+  private void handleFetchSignInMethodsForEmail(
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
     String email = arguments.get("email");
 
     firebaseAuth
-        .fetchProvidersForEmail(email)
-        .addOnCompleteListener(new ProvidersCompleteListener(result));
+        .fetchSignInMethodsForEmail(email)
+        .addOnCompleteListener(new GetSignInMethodsCompleteListener(result));
   }
 
   private void handleSendPasswordResetEmail(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
     String email = arguments.get("email");
 
     firebaseAuth
@@ -329,126 +324,97 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private void handleSendEmailVerification(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+      @SuppressWarnings("unused") MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth
         .getCurrentUser()
         .sendEmailVerification()
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleReload(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+  private void handleReload(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth
         .getCurrentUser()
         .reload()
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleSignInWithEmailAndPassword(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String email = arguments.get("email");
-    String password = arguments.get("password");
-
-    firebaseAuth
-        .signInWithEmailAndPassword(email, password)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleDelete(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+  private void handleDelete(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     firebaseAuth
         .getCurrentUser()
         .delete()
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleSignInWithGoogle(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String idToken = arguments.get("idToken");
-    String accessToken = arguments.get("accessToken");
-    AuthCredential credential = GoogleAuthProvider.getCredential(idToken, accessToken);
+  private AuthCredential getCredential(Map<String, Object> arguments) {
+    AuthCredential credential;
+    Map<String, String> data = (Map<String, String>) arguments.get("data");
+    switch ((String) arguments.get("provider")) {
+      case EmailAuthProvider.PROVIDER_ID:
+        {
+          String email = data.get("email");
+          String password = data.get("password");
+          credential = EmailAuthProvider.getCredential(email, password);
+          break;
+        }
+      case GoogleAuthProvider.PROVIDER_ID:
+        {
+          String idToken = data.get("idToken");
+          String accessToken = data.get("accessToken");
+          credential = GoogleAuthProvider.getCredential(idToken, accessToken);
+          break;
+        }
+      case FacebookAuthProvider.PROVIDER_ID:
+        {
+          String accessToken = data.get("accessToken");
+          credential = FacebookAuthProvider.getCredential(accessToken);
+          break;
+        }
+      case TwitterAuthProvider.PROVIDER_ID:
+        {
+          String authToken = data.get("authToken");
+          String authTokenSecret = data.get("authTokenSecret");
+          credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
+          break;
+        }
+      case GithubAuthProvider.PROVIDER_ID:
+        {
+          String token = data.get("token");
+          credential = GithubAuthProvider.getCredential(token);
+          break;
+        }
+      default:
+        {
+          credential = null;
+          break;
+        }
+    }
+    return credential;
+  }
+
+  private void handleSignInWithCredential(
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    AuthCredential credential = getCredential((Map<String, Object>) call.arguments());
     firebaseAuth
         .signInWithCredential(credential)
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
-  private void handleLinkWithGoogleCredential(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String idToken = arguments.get("idToken");
-    String accessToken = arguments.get("accessToken");
-    AuthCredential credential = GoogleAuthProvider.getCredential(idToken, accessToken);
+  private void handleReauthenticateWithCredential(
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    AuthCredential credential = getCredential((Map<String, Object>) call.arguments());
     firebaseAuth
         .getCurrentUser()
-        .linkWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
+        .reauthenticate(credential)
+        .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleLinkWithFacebookCredential(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String accessToken = arguments.get("accessToken");
-    AuthCredential credential = FacebookAuthProvider.getCredential(accessToken);
+  private void handleUnlinkFromProvider(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
+    final String provider = arguments.get("provider");
+
     firebaseAuth
         .getCurrentUser()
-        .linkWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleLinkWithTwitterCredential(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String authToken = arguments.get("authToken");
-    String authTokenSecret = arguments.get("authTokenSecret");
-    AuthCredential credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
-    firebaseAuth
-        .getCurrentUser()
-        .linkWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleLinkWithGithubCredential(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    String token = call.argument("token");
-    AuthCredential credential = GithubAuthProvider.getCredential(token);
-    firebaseAuth
-        .getCurrentUser()
-        .linkWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleSignInWithFacebook(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
-    String accessToken = arguments.get("accessToken");
-    AuthCredential credential = FacebookAuthProvider.getCredential(accessToken);
-    firebaseAuth
-        .signInWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleSignInWithTwitter(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    String authToken = call.argument("authToken");
-    String authTokenSecret = call.argument("authTokenSecret");
-    AuthCredential credential = TwitterAuthProvider.getCredential(authToken, authTokenSecret);
-    firebaseAuth
-        .signInWithCredential(credential)
-        .addOnCompleteListener(new SignInCompleteListener(result));
-  }
-
-  private void handleSignInWithGithub(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    String token = call.argument("token");
-    AuthCredential credential = GithubAuthProvider.getCredential(token);
-    firebaseAuth
-        .signInWithCredential(credential)
+        .unlink(provider)
         .addOnCompleteListener(new SignInCompleteListener(result));
   }
 
@@ -456,6 +422,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
     Map<String, String> arguments = call.arguments();
     String token = arguments.get("token");
+
     firebaseAuth
         .signInWithCustomToken(token)
         .addOnCompleteListener(new SignInCompleteListener(result));
@@ -467,9 +434,9 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private void handleGetToken(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, Boolean> arguments = (Map<String, Boolean>) call.arguments;
+    Map<String, Boolean> arguments = call.arguments();
     boolean refresh = arguments.get("refresh");
+
     firebaseAuth
         .getCurrentUser()
         .getIdToken(refresh)
@@ -480,35 +447,34 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
                   String idToken = task.getResult().getToken();
                   result.success(idToken);
                 } else {
-                  result.error(ERROR_REASON_EXCEPTION, task.getException().getMessage(), null);
+                  reportException(result, task.getException());
                 }
               }
             });
   }
 
-  private void handleUpdateEmail(MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+  private void handleUpdateEmail(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
+    final String email = arguments.get("email");
+
     firebaseAuth
         .getCurrentUser()
-        .updateEmail(arguments.get("email"))
+        .updateEmail(email)
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleUpdatePassword(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+  private void handleUpdatePassword(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
+    final String password = arguments.get("password");
+
     firebaseAuth
         .getCurrentUser()
-        .updatePassword(arguments.get("password"))
+        .updatePassword(password)
         .addOnCompleteListener(new TaskVoidCompleteListener(result));
   }
 
-  private void handleUpdateProfile(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+  private void handleUpdateProfile(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
 
     UserProfileChangeRequest.Builder builder = new UserProfileChangeRequest.Builder();
     if (arguments.containsKey("displayName")) {
@@ -525,7 +491,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private void handleStartListeningAuthState(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+      @SuppressWarnings("unused") MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     final int handle = nextHandle++;
     FirebaseAuth.AuthStateListener listener =
         new FirebaseAuth.AuthStateListener() {
@@ -535,7 +501,6 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
             Map<String, Object> userMap = mapFromUser(user);
             Map<String, Object> map = new HashMap<>();
             map.put("id", handle);
-
             if (userMap != null) {
               map.put("user", userMap);
             }
@@ -548,7 +513,7 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
   }
 
   private void handleStopListeningAuthState(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
+      MethodCall call, Result result, FirebaseAuth firebaseAuth) {
     Map<String, Integer> arguments = call.arguments();
     Integer id = arguments.get("id");
 
@@ -558,17 +523,16 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       authStateListeners.remove(id);
       result.success(null);
     } else {
-      result.error(
-          ERROR_REASON_EXCEPTION,
-          String.format("Listener with identifier '%d' not found.", id),
-          null);
+      reportException(
+          result,
+          new FirebaseAuthException(
+              "ERROR_LISTENER_NOT_FOUND",
+              String.format(Locale.US, "Listener with identifier '%d' not found.", id)));
     }
   }
 
-  private void handleSetLanguageCode(
-      MethodCall call, final Result result, FirebaseAuth firebaseAuth) {
-    @SuppressWarnings("unchecked")
-    Map<String, String> arguments = (Map<String, String>) call.arguments;
+  private void handleSetLanguageCode(MethodCall call, Result result, FirebaseAuth firebaseAuth) {
+    Map<String, String> arguments = call.arguments();
     String language = arguments.get("language");
 
     firebaseAuth.setLanguageCode(language);
@@ -584,9 +548,8 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
 
     @Override
     public void onComplete(@NonNull Task<AuthResult> task) {
-      if (!task.isSuccessful()) {
-        Exception e = task.getException();
-        result.error(ERROR_REASON_EXCEPTION, e.getMessage(), null);
+      if (!task.isSuccessful() || task.getResult() == null) {
+        reportException(result, task.getException());
       } else {
         FirebaseUser user = task.getResult().getUser();
         Map<String, Object> userMap = Collections.unmodifiableMap(mapFromUser(user));
@@ -605,28 +568,27 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
     @Override
     public void onComplete(@NonNull Task<Void> task) {
       if (!task.isSuccessful()) {
-        Exception e = task.getException();
-        result.error(ERROR_REASON_EXCEPTION, e.getMessage(), null);
+        reportException(result, task.getException());
       } else {
         result.success(null);
       }
     }
   }
 
-  private class ProvidersCompleteListener implements OnCompleteListener<ProviderQueryResult> {
+  private class GetSignInMethodsCompleteListener
+      implements OnCompleteListener<SignInMethodQueryResult> {
     private final Result result;
 
-    ProvidersCompleteListener(Result result) {
+    GetSignInMethodsCompleteListener(Result result) {
       this.result = result;
     }
 
     @Override
-    public void onComplete(@NonNull Task<ProviderQueryResult> task) {
-      if (!task.isSuccessful()) {
-        Exception e = task.getException();
-        result.error(ERROR_REASON_EXCEPTION, e.getMessage(), null);
+    public void onComplete(@NonNull Task<SignInMethodQueryResult> task) {
+      if (!task.isSuccessful() || task.getResult() == null) {
+        reportException(result, task.getException());
       } else {
-        List<String> providers = task.getResult().getProviders();
+        List<String> providers = task.getResult().getSignInMethods();
         result.success(providers);
       }
     }
@@ -670,6 +632,25 @@ public class FirebaseAuthPlugin implements MethodCallHandler {
       return Collections.unmodifiableMap(userMap);
     } else {
       return null;
+    }
+  }
+
+  private void reportException(Result result, @Nullable Exception exception) {
+    if (exception != null) {
+      if (exception instanceof FirebaseAuthException) {
+        final FirebaseAuthException authException = (FirebaseAuthException) exception;
+        result.error(authException.getErrorCode(), exception.getMessage(), null);
+      } else if (exception instanceof FirebaseApiNotAvailableException) {
+        result.error("ERROR_API_NOT_AVAILABLE", exception.getMessage(), null);
+      } else if (exception instanceof FirebaseTooManyRequestsException) {
+        result.error("ERROR_TOO_MANY_REQUESTS", exception.getMessage(), null);
+      } else if (exception instanceof FirebaseNetworkException) {
+        result.error("ERROR_NETWORK_REQUEST_FAILED", exception.getMessage(), null);
+      } else {
+        result.error(exception.getClass().getSimpleName(), exception.getMessage(), null);
+      }
+    } else {
+      result.error("ERROR_UNKNOWN", "An unknown error occurred.", null);
     }
   }
 }
