@@ -6,15 +6,17 @@
 
 #import "Firebase/Firebase.h"
 
-@interface NSError (FlutterError)
-@property(readonly, nonatomic) FlutterError *flutterError;
+@interface NSError (FIRAuthErrorCode)
+@property(readonly, nonatomic) NSString *firAuthErrorCode;
 @end
 
-@implementation NSError (FlutterError)
-- (FlutterError *)flutterError {
-  return [FlutterError errorWithCode:[NSString stringWithFormat:@"Error %d", (int)self.code]
-                             message:self.domain
-                             details:self.localizedDescription];
+@implementation NSError (FIRAuthErrorCode)
+- (NSString *)firAuthErrorCode {
+  NSString *code = [self userInfo][FIRAuthErrorNameKey];
+  if (code != nil) {
+    return code;
+  }
+  return [NSString stringWithFormat:@"ERROR_%d", (int)self.code];
 }
 @end
 
@@ -73,76 +75,50 @@ int nextHandle = 0;
         }];
   } else if ([@"signInAnonymously" isEqualToString:call.method]) {
     [[self getAuth:call.arguments]
-        signInAnonymouslyWithCompletion:^(FIRAuthDataResult *dataResult, NSError *error) {
-          [self sendResult:result forUser:dataResult.user error:error];
+        signInAnonymouslyWithCompletion:^(FIRAuthDataResult *authResult, NSError *error) {
+          [self sendResult:result forUser:authResult.user error:error];
         }];
-  } else if ([@"signInWithGoogle" isEqualToString:call.method]) {
-    NSString *idToken = call.arguments[@"idToken"];
-    NSString *accessToken = call.arguments[@"accessToken"];
-    FIRAuthCredential *credential = [FIRGoogleAuthProvider credentialWithIDToken:idToken
-                                                                     accessToken:accessToken];
-    [[self getAuth:call.arguments] signInWithCredential:credential
-                                             completion:^(FIRUser *user, NSError *error) {
-                                               [self sendResult:result forUser:user error:error];
-                                             }];
-  } else if ([@"signInWithFacebook" isEqualToString:call.method]) {
-    NSString *accessToken = call.arguments[@"accessToken"];
-    FIRAuthCredential *credential = [FIRFacebookAuthProvider credentialWithAccessToken:accessToken];
-    [[self getAuth:call.arguments] signInWithCredential:credential
-                                             completion:^(FIRUser *user, NSError *error) {
-                                               [self sendResult:result forUser:user error:error];
-                                             }];
-  } else if ([@"signInWithTwitter" isEqualToString:call.method]) {
-    NSString *authToken = call.arguments[@"authToken"];
-    NSString *authTokenSecret = call.arguments[@"authTokenSecret"];
-    FIRAuthCredential *credential = [FIRTwitterAuthProvider credentialWithToken:authToken
-                                                                         secret:authTokenSecret];
-    [[self getAuth:call.arguments] signInWithCredential:credential
-                                             completion:^(FIRUser *user, NSError *error) {
-                                               [self sendResult:result forUser:user error:error];
-                                             }];
-  } else if ([@"signInWithGithub" isEqualToString:call.method]) {
-    NSString *token = call.arguments[@"token"];
-    FIRAuthCredential *credential = [FIRGitHubAuthProvider credentialWithToken:token];
-    [[self getAuth:call.arguments] signInWithCredential:credential
-                                             completion:^(FIRUser *user, NSError *error) {
-                                               [self sendResult:result forUser:user error:error];
-                                             }];
+  } else if ([@"signInWithCredential" isEqualToString:call.method]) {
+    [[self getAuth:call.arguments]
+        signInAndRetrieveDataWithCredential:[self getCredential:call.arguments]
+                                 completion:^(FIRAuthDataResult *authResult, NSError *error) {
+                                   [self sendResult:result forUser:authResult.user error:error];
+                                 }];
   } else if ([@"createUserWithEmailAndPassword" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     NSString *password = call.arguments[@"password"];
     [[self getAuth:call.arguments]
         createUserWithEmail:email
                    password:password
-                 completion:^(FIRAuthDataResult *dataResult, NSError *error) {
-                   [self sendResult:result forUser:dataResult.user error:error];
+                 completion:^(FIRAuthDataResult *authResult, NSError *error) {
+                   [self sendResult:result forUser:authResult.user error:error];
                  }];
-  } else if ([@"fetchProvidersForEmail" isEqualToString:call.method]) {
+  } else if ([@"fetchSignInMethodsForEmail" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     [[self getAuth:call.arguments]
         fetchProvidersForEmail:email
                     completion:^(NSArray<NSString *> *providers, NSError *error) {
-                      [self sendResult:result forProviders:providers error:error];
+                      [self sendResult:result forObject:providers error:error];
                     }];
   } else if ([@"sendEmailVerification" isEqualToString:call.method]) {
     [[self getAuth:call.arguments].currentUser
         sendEmailVerificationWithCompletion:^(NSError *_Nullable error) {
-          [self sendResult:result forProviders:nil error:error];
+          [self sendResult:result forObject:nil error:error];
         }];
   } else if ([@"reload" isEqualToString:call.method]) {
     [[self getAuth:call.arguments].currentUser reloadWithCompletion:^(NSError *_Nullable error) {
-      [self sendResult:result forProviders:nil error:error];
+      [self sendResult:result forObject:nil error:error];
     }];
   } else if ([@"delete" isEqualToString:call.method]) {
     [[self getAuth:call.arguments].currentUser deleteWithCompletion:^(NSError *_Nullable error) {
-      [self sendResult:result forProviders:nil error:error];
+      [self sendResult:result forObject:nil error:error];
     }];
   } else if ([@"sendPasswordResetEmail" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     [[self getAuth:call.arguments] sendPasswordResetWithEmail:email
                                                    completion:^(NSError *error) {
                                                      [self sendResult:result
-                                                              forUser:nil
+                                                            forObject:nil
                                                                 error:error];
                                                    }];
   } else if ([@"signInWithEmailAndPassword" isEqualToString:call.method]) {
@@ -151,87 +127,57 @@ int nextHandle = 0;
     [[self getAuth:call.arguments]
         signInWithEmail:email
                password:password
-             completion:^(FIRAuthDataResult *dataResult, NSError *error) {
-               [self sendResult:result forUser:dataResult.user error:error];
+             completion:^(FIRAuthDataResult *authResult, NSError *error) {
+               [self sendResult:result forUser:authResult.user error:error];
              }];
   } else if ([@"signOut" isEqualToString:call.method]) {
     NSError *signOutError;
     BOOL status = [[self getAuth:call.arguments] signOut:&signOutError];
     if (!status) {
       NSLog(@"Error signing out: %@", signOutError);
-      [self sendResult:result forUser:nil error:signOutError];
+      [self sendResult:result forObject:nil error:signOutError];
     } else {
-      [self sendResult:result forUser:nil error:nil];
+      [self sendResult:result forObject:nil error:nil];
     }
   } else if ([@"getIdToken" isEqualToString:call.method]) {
     [[self getAuth:call.arguments].currentUser
         getIDTokenForcingRefresh:YES
                       completion:^(NSString *_Nullable token, NSError *_Nullable error) {
-                        result(error != nil ? error.flutterError : token);
+                        [self sendResult:result forObject:token error:error];
                       }];
-  } else if ([@"linkWithEmailAndPassword" isEqualToString:call.method]) {
-    NSString *email = call.arguments[@"email"];
-    NSString *password = call.arguments[@"password"];
-    FIRAuthCredential *credential = [FIREmailAuthProvider credentialWithEmail:email
-                                                                     password:password];
-    [[self getAuth:call.arguments].currentUser linkWithCredential:credential
-                                                       completion:^(FIRUser *user, NSError *error) {
-                                                         [self sendResult:result
-                                                                  forUser:user
-                                                                    error:error];
-                                                       }];
-  } else if ([@"linkWithGoogleCredential" isEqualToString:call.method]) {
-    NSString *idToken = call.arguments[@"idToken"];
-    NSString *accessToken = call.arguments[@"accessToken"];
-    FIRAuthCredential *credential = [FIRGoogleAuthProvider credentialWithIDToken:idToken
-                                                                     accessToken:accessToken];
-    [[self getAuth:call.arguments].currentUser linkWithCredential:credential
-                                                       completion:^(FIRUser *user, NSError *error) {
-                                                         [self sendResult:result
-                                                                  forUser:user
-                                                                    error:error];
-                                                       }];
-  } else if ([@"linkWithFacebookCredential" isEqualToString:call.method]) {
-    NSString *accessToken = call.arguments[@"accessToken"];
-    FIRAuthCredential *credential = [FIRFacebookAuthProvider credentialWithAccessToken:accessToken];
-    [[self getAuth:call.arguments].currentUser linkWithCredential:credential
-                                                       completion:^(FIRUser *user, NSError *error) {
-                                                         [self sendResult:result
-                                                                  forUser:user
-                                                                    error:error];
-                                                       }];
-  } else if ([@"linkWithTwitterCredential" isEqualToString:call.method]) {
-    NSString *authToken = call.arguments[@"authToken"];
-    NSString *authTokenSecret = call.arguments[@"authTokenSecret"];
-    FIRAuthCredential *credential = [FIRTwitterAuthProvider credentialWithToken:authToken
-                                                                         secret:authTokenSecret];
-    [[self getAuth:call.arguments].currentUser linkWithCredential:credential
-                                                       completion:^(FIRUser *user, NSError *error) {
-                                                         [self sendResult:result
-                                                                  forUser:user
-                                                                    error:error];
-                                                       }];
-  } else if ([@"linkWithGithubCredential" isEqualToString:call.method]) {
-    NSString *token = call.arguments[@"token"];
-    FIRAuthCredential *credential = [FIRGitHubAuthProvider credentialWithToken:token];
-    [[self getAuth:call.arguments].currentUser linkWithCredential:credential
-                                                       completion:^(FIRUser *user, NSError *error) {
-                                                         [self sendResult:result
-                                                                  forUser:user
-                                                                    error:error];
-                                                       }];
+  } else if ([@"reauthenticateWithCredential" isEqualToString:call.method]) {
+    [[self getAuth:call.arguments].currentUser
+        reauthenticateWithCredential:[self getCredential:call.arguments]
+                          completion:^(NSError *_Nullable error) {
+                            [self sendResult:result forObject:nil error:error];
+                          }];
+  } else if ([@"linkWithCredential" isEqualToString:call.method]) {
+    [[self getAuth:call.arguments].currentUser
+        linkWithCredential:[self getCredential:call.arguments]
+                completion:^(FIRUser *user, NSError *error) {
+                  [self sendResult:result forUser:user error:error];
+                }];
+  } else if ([@"unlinkFromProvider" isEqualToString:call.method]) {
+    NSString *provider = call.arguments[@"provider"];
+    [[self getAuth:call.arguments].currentUser
+        unlinkFromProvider:provider
+                completion:^(FIRUser *_Nullable user, NSError *_Nullable error) {
+                  [self sendResult:result forUser:user error:error];
+                }];
   } else if ([@"updateEmail" isEqualToString:call.method]) {
     NSString *email = call.arguments[@"email"];
     [[self getAuth:call.arguments].currentUser updateEmail:email
                                                 completion:^(NSError *error) {
-                                                  [self sendResult:result forUser:nil error:error];
+                                                  [self sendResult:result
+                                                         forObject:nil
+                                                             error:error];
                                                 }];
   } else if ([@"updatePassword" isEqualToString:call.method]) {
     NSString *password = call.arguments[@"password"];
     [[self getAuth:call.arguments].currentUser updatePassword:password
                                                    completion:^(NSError *error) {
                                                      [self sendResult:result
-                                                              forUser:nil
+                                                            forObject:nil
                                                                 error:error];
                                                    }];
   } else if ([@"updateProfile" isEqualToString:call.method]) {
@@ -244,20 +190,14 @@ int nextHandle = 0;
       changeRequest.photoURL = [NSURL URLWithString:call.arguments[@"photoUrl"]];
     }
     [changeRequest commitChangesWithCompletion:^(NSError *error) {
-      [self sendResult:result forUser:nil error:error];
+      [self sendResult:result forObject:nil error:error];
     }];
-  } else if ([@"updateEmail" isEqualToString:call.method]) {
-    NSString *toEmail = call.arguments[@"email"];
-    [[self getAuth:call.arguments].currentUser updateEmail:toEmail
-                                                completion:^(NSError *_Nullable error) {
-                                                  [self sendResult:result forUser:nil error:error];
-                                                }];
   } else if ([@"signInWithCustomToken" isEqualToString:call.method]) {
     NSString *token = call.arguments[@"token"];
     [[self getAuth:call.arguments]
         signInWithCustomToken:token
-                   completion:^(FIRAuthDataResult *dataResult, NSError *error) {
-                     [self sendResult:result forUser:dataResult.user error:error];
+                   completion:^(FIRAuthDataResult *authResult, NSError *error) {
+                     [self sendResult:result forUser:authResult.user error:error];
                    }];
 
   } else if ([@"startListeningAuthState" isEqualToString:call.method]) {
@@ -286,7 +226,7 @@ int nextHandle = 0;
       result(nil);
     } else {
       result([FlutterError
-          errorWithCode:@"not_found"
+          errorWithCode:@"ERROR_LISTENER_NOT_FOUND"
                 message:[NSString stringWithFormat:@"Listener with identifier '%d' not found.",
                                                    identifier.intValue]
                 details:nil]);
@@ -318,14 +258,15 @@ int nextHandle = 0;
     FIRPhoneAuthCredential *credential =
         [[FIRPhoneAuthProvider provider] credentialWithVerificationID:verificationId
                                                      verificationCode:smsCode];
-    [[self getAuth:call.arguments] signInWithCredential:credential
-                                             completion:^(FIRUser *user, NSError *error) {
-                                               [self sendResult:result forUser:user error:error];
-                                             }];
+    [[self getAuth:call.arguments]
+        signInWithCredential:credential
+                  completion:^(FIRUser *_Nullable user, NSError *_Nullable error) {
+                    [self sendResult:result forUser:user error:error];
+                  }];
   } else if ([@"setLanguageCode" isEqualToString:call.method]) {
     NSString *language = call.arguments[@"language"];
     [[self getAuth:call.arguments] setLanguageCode:language];
-    [self sendResult:result forUser:nil error:nil];
+    [self sendResult:result forObject:nil error:nil];
   } else {
     result(FlutterMethodNotImplemented);
   }
@@ -351,24 +292,20 @@ int nextHandle = 0;
 }
 
 - (void)sendResult:(FlutterResult)result forUser:(FIRUser *)user error:(NSError *)error {
-  if (error != nil) {
-    result(error.flutterError);
-  } else if (user == nil) {
-    result(nil);
-  } else {
-    result([self dictionaryFromUser:user]);
-  }
+  [self sendResult:result
+         forObject:(user != nil ? [self dictionaryFromUser:user] : nil)
+             error:error];
 }
 
-- (void)sendResult:(FlutterResult)result
-      forProviders:(NSArray<NSString *> *)providers
-             error:(NSError *)error {
+- (void)sendResult:(FlutterResult)result forObject:(NSObject *)object error:(NSError *)error {
   if (error != nil) {
-    result(error.flutterError);
-  } else if (providers == nil) {
+    result([FlutterError errorWithCode:error.firAuthErrorCode
+                               message:error.localizedDescription
+                               details:nil]);
+  } else if (object == nil) {
     result(nil);
   } else {
-    result(providers);
+    result(object);
   }
 }
 
@@ -385,5 +322,39 @@ int nextHandle = 0;
     errorCode = @"missingPhoneNumber";
   }
   return @{@"code" : errorCode, @"message" : error.localizedDescription};
+}
+
+- (FIRAuthCredential *)getCredential:(NSDictionary *)arguments {
+  NSString *provider = arguments[@"provider"];
+  NSDictionary *data = arguments[@"data"];
+  FIRAuthCredential *credential;
+  if ([FIREmailAuthProviderID isEqualToString:provider]) {
+    NSString *email = data[@"email"];
+    NSString *password = data[@"password"];
+    credential = [FIREmailAuthProvider credentialWithEmail:email password:password];
+  } else if ([FIRGoogleAuthProviderID isEqualToString:provider]) {
+    NSString *idToken = data[@"idToken"];
+    NSString *accessToken = data[@"accessToken"];
+    credential = [FIRGoogleAuthProvider credentialWithIDToken:idToken accessToken:accessToken];
+  } else if ([FIRFacebookAuthProviderID isEqualToString:provider]) {
+    NSString *accessToken = data[@"accessToken"];
+    credential = [FIRFacebookAuthProvider credentialWithAccessToken:accessToken];
+  } else if ([FIRTwitterAuthProviderID isEqualToString:provider]) {
+    NSString *authToken = data[@"authToken"];
+    NSString *authTokenSecret = data[@"authTokenSecret"];
+    credential = [FIRTwitterAuthProvider credentialWithToken:authToken secret:authTokenSecret];
+  } else if ([FIRGitHubAuthProviderID isEqualToString:provider]) {
+    NSString *token = data[@"token"];
+    credential = [FIRGitHubAuthProvider credentialWithToken:token];
+  } else if ([FIRPhoneAuthProviderID isEqualToString:provider]) {
+    NSString *verificationId = data[@"verificationId"];
+    NSString *smsCode = data[@"smsCode"];
+    credential = [[FIRPhoneAuthProvider providerWithAuth:[self getAuth:arguments]]
+        credentialWithVerificationID:verificationId
+                    verificationCode:smsCode];
+  } else {
+    NSLog(@"Support for an auth provider with identifier '%@' is not implemented.", provider);
+  }
+  return credential;
 }
 @end
