@@ -2,6 +2,7 @@ package io.flutter.plugins.firebasemlvision;
 
 import android.net.Uri;
 import com.google.firebase.ml.vision.common.FirebaseVisionImage;
+import com.google.firebase.ml.vision.common.FirebaseVisionImageMetadata;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
 import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
@@ -29,55 +30,74 @@ public class FirebaseMlVisionPlugin implements MethodCallHandler {
   @Override
   public void onMethodCall(MethodCall call, Result result) {
     Map<String, Object> options = call.argument("options");
+
     FirebaseVisionImage image;
+    Map<String, Object> imageData = call.arguments();
+    try {
+      image = dataToVisionImage(imageData);
+    } catch (IOException exception) {
+      result.error("MLVisionDetectorIOError", exception.getLocalizedMessage(), null);
+      return;
+    }
+
     switch (call.method) {
       case "BarcodeDetector#detectInImage":
-        try {
-          image = filePathToVisionImage((String) call.argument("path"));
-          BarcodeDetector.instance.handleDetection(image, options, result);
-        } catch (IOException e) {
-          result.error("barcodeDetectorIOError", e.getLocalizedMessage(), null);
-        }
+        BarcodeDetector.instance.handleDetection(image, options, result);
         break;
       case "FaceDetector#detectInImage":
-        try {
-          image = filePathToVisionImage((String) call.argument("path"));
-          FaceDetector.instance.handleDetection(image, options, result);
-        } catch (IOException e) {
-          result.error("faceDetectorIOError", e.getLocalizedMessage(), null);
-        }
+        FaceDetector.instance.handleDetection(image, options, result);
         break;
       case "LabelDetector#detectInImage":
-        try {
-          image = filePathToVisionImage((String) call.argument("path"));
-          LabelDetector.instance.handleDetection(image, options, result);
-        } catch (IOException e) {
-          result.error("labelDetectorIOError", e.getLocalizedMessage(), null);
-        }
+        LabelDetector.instance.handleDetection(image, options, result);
         break;
       case "CloudLabelDetector#detectInImage":
-        try {
-          image = filePathToVisionImage((String) call.argument("path"));
-          CloudLabelDetector.instance.handleDetection(image, options, result);
-        } catch (IOException e) {
-          result.error("cloudLabelDetectorIOError", e.getLocalizedMessage(), null);
-        }
+        CloudLabelDetector.instance.handleDetection(image, options, result);
         break;
       case "TextRecognizer#processImage":
-        try {
-          image = filePathToVisionImage((String) call.argument("path"));
-          TextRecognizer.instance.handleDetection(image, options, result);
-        } catch (IOException e) {
-          result.error("textRecognizerIOError", e.getLocalizedMessage(), null);
-        }
+        TextRecognizer.instance.handleDetection(image, options, result);
         break;
       default:
         result.notImplemented();
     }
   }
 
-  private FirebaseVisionImage filePathToVisionImage(String path) throws IOException {
-    File file = new File(path);
-    return FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
+  private FirebaseVisionImage dataToVisionImage(Map<String, Object> imageData) throws IOException {
+    String imageType = (String) imageData.get("type");
+
+    switch (imageType) {
+      case "file":
+        File file = new File((String) imageData.get("path"));
+        return FirebaseVisionImage.fromFilePath(registrar.context(), Uri.fromFile(file));
+      case "bytes":
+        @SuppressWarnings("unchecked")
+        Map<String, Object> metadataData = (Map<String, Object>) imageData.get("metadata");
+
+        FirebaseVisionImageMetadata metadata =
+            new FirebaseVisionImageMetadata.Builder()
+                .setWidth((int) (double) metadataData.get("width"))
+                .setHeight((int) (double) metadataData.get("height"))
+                .setFormat(FirebaseVisionImageMetadata.IMAGE_FORMAT_NV21)
+                .setRotation(getRotation((int) metadataData.get("rotation")))
+                .build();
+
+        return FirebaseVisionImage.fromByteArray((byte[]) imageData.get("bytes"), metadata);
+      default:
+        throw new IllegalArgumentException(String.format("No image type for: %s", imageType));
+    }
+  }
+
+  private int getRotation(int rotation) {
+    switch (rotation) {
+      case 0:
+        return FirebaseVisionImageMetadata.ROTATION_0;
+      case 90:
+        return FirebaseVisionImageMetadata.ROTATION_90;
+      case 180:
+        return FirebaseVisionImageMetadata.ROTATION_180;
+      case 270:
+        return FirebaseVisionImageMetadata.ROTATION_270;
+      default:
+        throw new IllegalArgumentException(String.format("No rotation for: %d", rotation));
+    }
   }
 }
