@@ -11,13 +11,15 @@ import static io.flutter.plugins.googlemaps.GoogleMapsPlugin.RESUMED;
 import static io.flutter.plugins.googlemaps.GoogleMapsPlugin.STARTED;
 import static io.flutter.plugins.googlemaps.GoogleMapsPlugin.STOPPED;
 
+import android.Manifest;
 import android.app.Activity;
 import android.app.Application;
 import android.content.Context;
+import android.content.pm.PackageManager;
 import android.os.Bundle;
+import android.util.Log;
 import android.view.View;
 import com.google.android.gms.maps.CameraUpdate;
-import com.google.android.gms.maps.CameraUpdateFactory;
 import com.google.android.gms.maps.GoogleMap;
 import com.google.android.gms.maps.GoogleMapOptions;
 import com.google.android.gms.maps.MapView;
@@ -48,6 +50,7 @@ final class GoogleMapController
         OnMapReadyCallback,
         OnMarkerTappedListener,
         PlatformView {
+  private static final String TAG = "GoogleMapController";
   private final int id;
   private final AtomicInteger activityState;
   private final MethodChannel methodChannel;
@@ -56,10 +59,12 @@ final class GoogleMapController
   private final Map<String, MarkerController> markers;
   private GoogleMap googleMap;
   private boolean trackCameraPosition = false;
+  private boolean myLocationEnabled = false;
   private boolean disposed = false;
   private final float density;
   private MethodChannel.Result mapReadyResult;
   private final int registrarActivityHashCode;
+  private final Context context;
 
   GoogleMapController(
       int id,
@@ -68,6 +73,7 @@ final class GoogleMapController
       PluginRegistry.Registrar registrar,
       GoogleMapOptions options) {
     this.id = id;
+    this.context = context;
     this.activityState = activityState;
     this.registrar = registrar;
     this.mapView = new MapView(context, options);
@@ -171,6 +177,7 @@ final class GoogleMapController
     googleMap.setOnCameraMoveListener(this);
     googleMap.setOnCameraIdleListener(this);
     googleMap.setOnMarkerClickListener(this);
+    updateMyLocationEnabled();
   }
 
   @Override
@@ -345,11 +352,6 @@ final class GoogleMapController
   // GoogleMapOptionsSink methods
 
   @Override
-  public void setCameraPosition(CameraPosition position) {
-    googleMap.moveCamera(CameraUpdateFactory.newCameraPosition(position));
-  }
-
-  @Override
   public void setCameraTargetBounds(LatLngBounds bounds) {
     googleMap.setLatLngBoundsForCameraTarget(bounds);
   }
@@ -398,5 +400,41 @@ final class GoogleMapController
   @Override
   public void setZoomGesturesEnabled(boolean zoomGesturesEnabled) {
     googleMap.getUiSettings().setZoomGesturesEnabled(zoomGesturesEnabled);
+  }
+
+  @Override
+  public void setMyLocationEnabled(boolean myLocationEnabled) {
+    if (this.myLocationEnabled == myLocationEnabled) {
+      return;
+    }
+    this.myLocationEnabled = myLocationEnabled;
+    if (googleMap != null) {
+      updateMyLocationEnabled();
+    }
+  }
+
+  private void updateMyLocationEnabled() {
+    if (hasLocationPermission()) {
+      googleMap.setMyLocationEnabled(myLocationEnabled);
+    } else {
+      // TODO(amirh): Make the options update fail.
+      // https://github.com/flutter/flutter/issues/24327
+      Log.e(TAG, "Cannot enable MyLocation layer as location permissions are not granted");
+    }
+  }
+
+  private boolean hasLocationPermission() {
+    return checkSelfPermission(Manifest.permission.ACCESS_FINE_LOCATION)
+            == PackageManager.PERMISSION_GRANTED
+        || checkSelfPermission(Manifest.permission.ACCESS_COARSE_LOCATION)
+            == PackageManager.PERMISSION_GRANTED;
+  }
+
+  private int checkSelfPermission(String permission) {
+    if (permission == null) {
+      throw new IllegalArgumentException("permission is null");
+    }
+    return context.checkPermission(
+        permission, android.os.Process.myPid(), android.os.Process.myUid());
   }
 }
