@@ -3,9 +3,11 @@ package io.flutter.plugins.googlemaps;
 import static io.flutter.plugin.common.PluginRegistry.Registrar;
 
 import android.content.Context;
+import com.google.android.gms.maps.model.CameraPosition;
 import io.flutter.plugin.common.StandardMessageCodec;
 import io.flutter.plugin.platform.PlatformView;
 import io.flutter.plugin.platform.PlatformViewFactory;
+import java.util.Map;
 import java.util.concurrent.atomic.AtomicInteger;
 
 public class GoogleMapFactory extends PlatformViewFactory {
@@ -20,9 +22,15 @@ public class GoogleMapFactory extends PlatformViewFactory {
   }
 
   @Override
-  public PlatformView create(Context context, int id, Object params) {
+  public PlatformView create(Context context, int id, Object args) {
+    Map<String, Object> params = (Map<String, Object>) args;
     final GoogleMapBuilder builder = new GoogleMapBuilder();
-    Convert.interpretGoogleMapOptions(params, builder);
+
+    Convert.interpretGoogleMapOptions(params.get("options"), builder);
+    if (params.containsKey("initialCameraPosition")) {
+      CameraPosition position = Convert.toCameraPosition(params.get("initialCameraPosition"));
+      builder.setInitialCameraPosition(position);
+    }
     return builder.build(id, context, mActivityState, mPluginRegistrar);
   }
 }
