@@ -175,11 +175,9 @@ class PlaceMarkerBodyState extends State<PlaceMarkerBody> {
             height: 200.0,
             child: GoogleMap(
               onMapCreated: _onMapCreated,
-              options: GoogleMapOptions(
-                cameraPosition: const CameraPosition(
-                  target: LatLng(-33.852, 151.211),
-                  zoom: 11.0,
-                ),
+              initialCameraPosition: const CameraPosition(
+                target: LatLng(-33.852, 151.211),
+                zoom: 11.0,
               ),
             ),
           ),
