@@ -4,8 +4,6 @@
 
 #import "GoogleMapController.h"
 
-static uint64_t _nextMapId = 0;
-
 #pragma mark - Conversion of JSON-like values sent via platform channels. Forward declarations.
 
 static id positionToJson(GMSCameraPosition* position);
@@ -60,12 +58,11 @@ static void interpretMarkerOptions(id json, id<FLTGoogleMapMarkerOptionsSink> si
   if ([super init]) {
     _viewId = viewId;
 
-    NSDictionary* options = args[@"options"];
-    GMSCameraPosition* camera = toOptionalCameraPosition(options[@"cameraPosition"]);
+    GMSCameraPosition* camera = toOptionalCameraPosition(args[@"initialCameraPosition"]);
     _mapView = [GMSMapView mapWithFrame:frame camera:camera];
     _markers = [NSMutableDictionary dictionaryWithCapacity:1];
     _trackCameraPosition = NO;
-    interpretMapOptions(options, self);
+    interpretMapOptions(args[@"options"], self);
     NSString* channelName =
         [NSString stringWithFormat:@"plugins.flutter.io/google_maps_%lld", viewId];
     _channel = [FlutterMethodChannel methodChannelWithName:channelName
@@ -76,6 +73,8 @@ static void interpretMarkerOptions(id json, id<FLTGoogleMapMarkerOptionsSink> si
         [weakSelf onMethodCall:call result:result];
       }
     }];
+    _mapView.delegate = weakSelf;
+    _registrar = registrar;
   }
   return self;
 }
@@ -117,17 +116,6 @@ static void interpretMarkerOptions(id json, id<FLTGoogleMapMarkerOptionsSink> si
   } else {
     result(FlutterMethodNotImplemented);
   }
-}
-
-- (void)addToView:(UIView*)view {
-  _mapView.hidden = YES;
-  _mapView.delegate = self;
-  [view addSubview:_mapView];
-}
-
-- (void)removeFromView {
-  [_mapView removeFromSuperview];
-  _mapView.delegate = nil;
 }
 
 - (void)showAtX:(CGFloat)x Y:(CGFloat)y {
@@ -217,32 +205,38 @@ static void interpretMarkerOptions(id json, id<FLTGoogleMapMarkerOptionsSink> si
   _mapView.settings.zoomGestures = enabled;
 }
 
+- (void)setMyLocationEnabled:(BOOL)enabled {
+  _mapView.myLocationEnabled = enabled;
+  _mapView.settings.myLocationButton = enabled;
+}
+
 #pragma mark - GMSMapViewDelegate methods
 
 - (void)mapView:(GMSMapView*)mapView willMove:(BOOL)gesture {
-  [_delegate onCameraMoveStartedOnMap:_mapId gesture:gesture];
+  [_channel invokeMethod:@"camera#onMoveStarted" arguments:@{@"isGesture" : @(gesture)}];
 }
 
 - (void)mapView:(GMSMapView*)mapView didChangeCameraPosition:(GMSCameraPosition*)position {
   if (_trackCameraPosition) {
-    [_delegate onCameraMoveOnMap:_mapId cameraPosition:position];
+    [_channel invokeMethod:@"camera#onMove" arguments:@{@"position" : positionToJson(position)}];
   }
 }
 
 - (void)mapView:(GMSMapView*)mapView idleAtCameraPosition:(GMSCameraPosition*)position {
-  [_delegate onCameraIdleOnMap:_mapId];
+  [_channel invokeMethod:@"camera#onIdle" arguments:@{}];
 }
 
 - (BOOL)mapView:(GMSMapView*)mapView didTapMarker:(GMSMarker*)marker {
   NSString* markerId = marker.userData[0];
-  [_delegate onMarkerTappedOnMap:_mapId marker:markerId];
+  [_channel invokeMethod:@"marker#onTap" arguments:@{@"marker" : markerId}];
   return [marker.userData[1] boolValue];
 }
 
-- (void)mapView:(GMSMapView*)mapView didTapInfoWindow:(GMSMarker*)marker {
+- (void)mapView:(GMSMapView*)mapView didTapInfoWindowOfMarker:(GMSMarker*)marker {
   NSString* markerId = marker.userData[0];
-  [_delegate onInfoWindowTappedOnMap:_mapId marker:markerId];
+  [_channel invokeMethod:@"infoWindow#onTap" arguments:@{@"marker" : markerId}];
 }
+
 @end
 
 #pragma mark - Implementations of JSON conversion functions.
@@ -352,10 +346,6 @@ static GMSCameraUpdate* toCameraUpdate(id json) {
 
 static void interpretMapOptions(id json, id<FLTGoogleMapOptionsSink> sink) {
   NSDictionary* data = json;
-  id cameraPosition = data[@"cameraPosition"];
-  if (cameraPosition) {
-    [sink setCamera:toCameraPosition(cameraPosition)];
-  }
   id cameraTargetBounds = data[@"cameraTargetBounds"];
   if (cameraTargetBounds) {
     [sink setCameraTargetBounds:toOptionalBounds(cameraTargetBounds)];
@@ -394,6 +384,10 @@ static void interpretMapOptions(id json, id<FLTGoogleMapOptionsSink> sink) {
   id zoomGesturesEnabled = data[@"zoomGesturesEnabled"];
   if (zoomGesturesEnabled) {
     [sink setZoomGesturesEnabled:toBool(zoomGesturesEnabled)];
+  }
+  id myLocationEnabled = data[@"myLocationEnabled"];
+  if (myLocationEnabled) {
+    [sink setMyLocationEnabled:toBool(myLocationEnabled)];
   }
 }
 
