@@ -18,28 +18,21 @@ part of google_maps_flutter;
 /// Marker tap events can be received by adding callbacks to [onMarkerTapped].
 class GoogleMapController extends ChangeNotifier {
   GoogleMapController._(
-      this._id, GoogleMapOptions options, MethodChannel channel)
+      this._id, MethodChannel channel, CameraPosition initialCameraPosition)
       : assert(_id != null),
-        assert(options != null),
-        assert(options.cameraPosition != null),
         assert(channel != null),
         _channel = channel {
-    if (options.trackCameraPosition) {
-      _cameraPosition = options.cameraPosition;
-    }
+    _cameraPosition = initialCameraPosition;
     _channel.setMethodCallHandler(_handleMethodCall);
-    _options = GoogleMapOptions.defaultOptions.copyWith(options);
   }
 
   static Future<GoogleMapController> init(
-      int id, GoogleMapOptions options) async {
+      int id, CameraPosition initialCameraPosition) async {
     assert(id != null);
-    assert(options != null);
-    assert(options.cameraPosition != null);
     final MethodChannel channel =
         MethodChannel('plugins.flutter.io/google_maps_$id');
     await channel.invokeMethod('map#waitForMap');
-    return GoogleMapController._(id, options, channel);
+    return GoogleMapController._(id, channel, initialCameraPosition);
   }
 
   final MethodChannel _channel;
@@ -50,11 +43,6 @@ class GoogleMapController extends ChangeNotifier {
   /// Callbacks to receive tap events for info windows on markers
   final ArgumentCallbacks<Marker> onInfoWindowTapped =
       ArgumentCallbacks<Marker>();
-
-  /// The configuration options most recently applied via controller
-  /// initialization or [updateMapOptions].
-  GoogleMapOptions get options => _options;
-  GoogleMapOptions _options;
 
   /// The current set of markers on this map.
   ///
@@ -67,8 +55,7 @@ class GoogleMapController extends ChangeNotifier {
   bool _isCameraMoving = false;
 
   /// Returns the most recent camera position reported by the platform side.
-  /// Will be null, if camera position tracking is not enabled via
-  /// [GoogleMapOptions].
+  /// Will be null, if [GoogleMap.trackCameraPosition] is false.
   CameraPosition get cameraPosition => _cameraPosition;
   CameraPosition _cameraPosition;
 
@@ -96,7 +83,7 @@ class GoogleMapController extends ChangeNotifier {
         notifyListeners();
         break;
       case 'camera#onMove':
-        _cameraPosition = CameraPosition._fromJson(call.arguments['position']);
+        _cameraPosition = CameraPosition.fromMap(call.arguments['position']);
         notifyListeners();
         break;
       case 'camera#onIdle':
@@ -114,16 +101,15 @@ class GoogleMapController extends ChangeNotifier {
   /// platform side.
   ///
   /// The returned [Future] completes after listeners have been notified.
-  Future<void> updateMapOptions(GoogleMapOptions changes) async {
-    assert(changes != null);
+  Future<void> _updateMapOptions(Map<String, dynamic> optionsUpdate) async {
+    assert(optionsUpdate != null);
     final dynamic json = await _channel.invokeMethod(
       'map#update',
       <String, dynamic>{
-        'options': changes._toJson(),
+        'options': optionsUpdate,
       },
     );
-    _options = _options.copyWith(changes);
-    _cameraPosition = CameraPosition._fromJson(json);
+    _cameraPosition = CameraPosition.fromMap(json);
     notifyListeners();
   }
 
