@@ -42,7 +42,23 @@ class CameraTargetBounds {
   /// Unbounded camera target.
   static const CameraTargetBounds unbounded = CameraTargetBounds(null);
 
-  dynamic _toJson() => <dynamic>[bounds?._toJson()];
+  dynamic _toJson() => <dynamic>[bounds?._toList()];
+
+  @override
+  bool operator ==(dynamic other) {
+    if (identical(this, other)) return true;
+    if (runtimeType != other.runtimeType) return false;
+    final CameraTargetBounds typedOther = other;
+    return bounds == typedOther.bounds;
+  }
+
+  @override
+  int get hashCode => bounds.hashCode;
+
+  @override
+  String toString() {
+    return 'CameraTargetBounds(bounds: $bounds)';
+  }
 }
 
 /// Preferred bounds for map camera zoom level.
@@ -64,136 +80,20 @@ class MinMaxZoomPreference {
       MinMaxZoomPreference(null, null);
 
   dynamic _toJson() => <dynamic>[minZoom, maxZoom];
-}
 
-/// Configuration options for the GoogleMaps user interface.
-///
-/// When used to change configuration, null values will be interpreted as
-/// "do not change this configuration option".
-class GoogleMapOptions {
-  /// Creates a set of map user interface configuration options.
-  ///
-  /// By default, every non-specified field is null, meaning no desire to change
-  /// user interface defaults or current configuration.
-  GoogleMapOptions({
-    this.cameraPosition,
-    this.compassEnabled,
-    this.cameraTargetBounds,
-    this.mapType,
-    this.minMaxZoomPreference,
-    this.rotateGesturesEnabled,
-    this.scrollGesturesEnabled,
-    this.tiltGesturesEnabled,
-    this.trackCameraPosition,
-    this.zoomGesturesEnabled,
-  });
-
-  /// The desired position of the map camera.
-  ///
-  /// This field is used to indicate initial camera position and to update that
-  /// position programmatically along with other changes to the map user
-  /// interface. It does not track the camera position through animations or
-  /// reflect movements caused by user touch events.
-  final CameraPosition cameraPosition;
-
-  /// True if the map should show a compass when rotated.
-  final bool compassEnabled;
-
-  /// Geographical bounding box for the camera target.
-  final CameraTargetBounds cameraTargetBounds;
-
-  /// Type of map tiles to be rendered.
-  final MapType mapType;
-
-  /// Preferred bounds for the camera zoom level.
-  ///
-  /// Actual bounds depend on map data and device.
-  final MinMaxZoomPreference minMaxZoomPreference;
-
-  /// True if the map view should respond to rotate gestures.
-  final bool rotateGesturesEnabled;
-
-  /// True if the map view should respond to scroll gestures.
-  final bool scrollGesturesEnabled;
-
-  /// True if the map view should respond to tilt gestures.
-  final bool tiltGesturesEnabled;
-
-  /// True if the map view should relay camera move events to Flutter.
-  final bool trackCameraPosition;
-
-  /// True if the map view should respond to zoom gestures.
-  final bool zoomGesturesEnabled;
-
-  /// Default user interface options.
-  ///
-  /// Specifies a map view that
-  /// * displays a compass when rotated; [compassEnabled] is true
-  /// * positions the camera at 0,0; [cameraPosition] has target `LatLng(0.0, 0.0)`
-  /// * does not bound the camera target; [cameraTargetBounds] is `CameraTargetBounds.unbounded`
-  /// * uses normal map tiles; [mapType] is `MapType.normal`
-  /// * does not bound zooming; [minMaxZoomPreference] is `MinMaxZoomPreference.unbounded`
-  /// * responds to rotate gestures; [rotateGesturesEnabled] is true
-  /// * responds to scroll gestures; [scrollGesturesEnabled] is true
-  /// * responds to tilt gestures; [tiltGesturesEnabled] is true
-  /// * is silent about camera movement; [trackCameraPosition] is false
-  /// * responds to zoom gestures; [zoomGesturesEnabled] is true
-  static final GoogleMapOptions defaultOptions = GoogleMapOptions(
-    compassEnabled: true,
-    cameraPosition: const CameraPosition(target: LatLng(0.0, 0.0)),
-    cameraTargetBounds: CameraTargetBounds.unbounded,
-    mapType: MapType.normal,
-    minMaxZoomPreference: MinMaxZoomPreference.unbounded,
-    rotateGesturesEnabled: true,
-    scrollGesturesEnabled: true,
-    tiltGesturesEnabled: true,
-    trackCameraPosition: false,
-    zoomGesturesEnabled: true,
-  );
-
-  /// Creates a new options object whose values are the same as this instance,
-  /// unless overwritten by the specified [changes].
-  ///
-  /// Returns this instance, if [changes] is null.
-  GoogleMapOptions copyWith(GoogleMapOptions change) {
-    if (change == null) {
-      return this;
-    }
-    return GoogleMapOptions(
-      cameraPosition: change.cameraPosition ?? cameraPosition,
-      compassEnabled: change.compassEnabled ?? compassEnabled,
-      cameraTargetBounds: change.cameraTargetBounds ?? cameraTargetBounds,
-      mapType: change.mapType ?? mapType,
-      minMaxZoomPreference: change.minMaxZoomPreference ?? minMaxZoomPreference,
-      rotateGesturesEnabled:
-          change.rotateGesturesEnabled ?? rotateGesturesEnabled,
-      scrollGesturesEnabled:
-          change.scrollGesturesEnabled ?? scrollGesturesEnabled,
-      tiltGesturesEnabled: change.tiltGesturesEnabled ?? tiltGesturesEnabled,
-      trackCameraPosition: change.trackCameraPosition ?? trackCameraPosition,
-      zoomGesturesEnabled: change.zoomGesturesEnabled ?? zoomGesturesEnabled,
-    );
+  @override
+  bool operator ==(dynamic other) {
+    if (identical(this, other)) return true;
+    if (runtimeType != other.runtimeType) return false;
+    final MinMaxZoomPreference typedOther = other;
+    return minZoom == typedOther.minZoom && maxZoom == typedOther.maxZoom;
   }
 
-  dynamic _toJson() {
-    final Map<String, dynamic> json = <String, dynamic>{};
+  @override
+  int get hashCode => hashValues(minZoom, maxZoom);
 
-    void addIfPresent(String fieldName, dynamic value) {
-      if (value != null) {
-        json[fieldName] = value;
-      }
-    }
-
-    addIfPresent('cameraPosition', cameraPosition?._toJson());
-    addIfPresent('compassEnabled', compassEnabled);
-    addIfPresent('cameraTargetBounds', cameraTargetBounds?._toJson());
-    addIfPresent('mapType', mapType?.index);
-    addIfPresent('minMaxZoomPreference', minMaxZoomPreference?._toJson());
-    addIfPresent('rotateGesturesEnabled', rotateGesturesEnabled);
-    addIfPresent('scrollGesturesEnabled', scrollGesturesEnabled);
-    addIfPresent('tiltGesturesEnabled', tiltGesturesEnabled);
-    addIfPresent('trackCameraPosition', trackCameraPosition);
-    addIfPresent('zoomGesturesEnabled', zoomGesturesEnabled);
-    return json;
+  @override
+  String toString() {
+    return 'MinMaxZoomPreference(minZoom: $minZoom, maxZoom: $maxZoom)';
   }
 }
