@@ -1,14 +1,13 @@
 package io.flutter.plugins.webviewflutter;
 
-import static io.flutter.plugin.common.MethodChannel.MethodCallHandler;
-import static io.flutter.plugin.common.MethodChannel.Result;
-
 import android.content.Context;
 import android.view.View;
 import android.webkit.WebView;
 import io.flutter.plugin.common.BinaryMessenger;
 import io.flutter.plugin.common.MethodCall;
 import io.flutter.plugin.common.MethodChannel;
+import io.flutter.plugin.common.MethodChannel.MethodCallHandler;
+import io.flutter.plugin.common.MethodChannel.Result;
 import io.flutter.plugin.platform.PlatformView;
 import java.util.Map;
 
@@ -42,6 +41,27 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
       case "updateSettings":
         updateSettings(methodCall, result);
         break;
+      case "canGoBack":
+        canGoBack(methodCall, result);
+        break;
+      case "canGoForward":
+        canGoForward(methodCall, result);
+        break;
+      case "goBack":
+        goBack(methodCall, result);
+        break;
+      case "goForward":
+        goForward(methodCall, result);
+        break;
+      case "reload":
+        reload(methodCall, result);
+        break;
+      case "currentUrl":
+        currentUrl(methodCall, result);
+        break;
+      case "evaluateJavascript":
+        evaluateJavaScript(methodCall, result);
+        break;
       default:
         result.notImplemented();
     }
@@ -53,10 +73,56 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
     result.success(null);
   }
 
+  private void canGoBack(MethodCall methodCall, Result result) {
+    result.success(webView.canGoBack());
+  }
+
+  private void canGoForward(MethodCall methodCall, Result result) {
+    result.success(webView.canGoForward());
+  }
+
+  private void goBack(MethodCall methodCall, Result result) {
+    if (webView.canGoBack()) {
+      webView.goBack();
+    }
+    result.success(null);
+  }
+
+  private void goForward(MethodCall methodCall, Result result) {
+    if (webView.canGoForward()) {
+      webView.goForward();
+    }
+    result.success(null);
+  }
+
+  private void reload(MethodCall methodCall, Result result) {
+    webView.reload();
+    result.success(null);
+  }
+
+  private void currentUrl(MethodCall methodCall, Result result) {
+    result.success(webView.getUrl());
+  }
+
   @SuppressWarnings("unchecked")
   private void updateSettings(MethodCall methodCall, Result result) {
     applySettings((Map<String, Object>) methodCall.arguments);
     result.success(null);
+  }
+
+  private void evaluateJavaScript(MethodCall methodCall, final Result result) {
+    String jsString = (String) methodCall.arguments;
+    if (jsString == null) {
+      throw new UnsupportedOperationException("JavaScript string cannot be null");
+    }
+    webView.evaluateJavascript(
+        jsString,
+        new android.webkit.ValueCallback<String>() {
+          @Override
+          public void onReceiveValue(String value) {
+            result.success(value);
+          }
+        });
   }
 
   private void applySettings(Map<String, Object> settings) {
@@ -76,11 +142,11 @@ public class FlutterWebView implements PlatformView, MethodCallHandler {
       case 0: // disabled
         webView.getSettings().setJavaScriptEnabled(false);
         break;
-      case 1: //unrestricted
+      case 1: // unrestricted
         webView.getSettings().setJavaScriptEnabled(true);
         break;
       default:
-        throw new IllegalArgumentException("Trying to set unknown Javascript mode: " + mode);
+        throw new IllegalArgumentException("Trying to set unknown JavaScript mode: " + mode);
     }
   }
 
