@@ -2,8 +2,8 @@ package io.flutter.plugins.webviewflutter;
 
 import io.flutter.plugin.common.PluginRegistry.Registrar;
 
-/** WebviewFlutterPlugin */
-public class WebviewFlutterPlugin {
+/** WebViewFlutterPlugin */
+public class WebViewFlutterPlugin {
   /** Plugin registration. */
   public static void registerWith(Registrar registrar) {
     registrar
